Separate components a bit better
